refactor: move VectorSettings schemas to his own file and unify canonical VectorSetting schema adding dataset_id
