+++ src/argilla/server/contexts/datasets.py
@@ -54,9 +54,6 @@ from argilla.server.models.suggestions import SuggestionCreateWithRecordId
 from argilla.server.schemas.v1.datasets import (
     DatasetCreate,
 )
-from argilla.server.schemas.v1.datasets import (
-    VectorSettings as VectorSettingsSchema,
-)
 from argilla.server.schemas.v1.fields import FieldCreate
 from argilla.server.schemas.v1.metadata_properties import MetadataPropertyCreate, MetadataPropertyUpdate
 from argilla.server.schemas.v1.questions import QuestionCreate
@@ -74,6 +71,12 @@ from argilla.server.schemas.v1.responses import (
     ResponseValueCreate,
     ResponseValueUpdate,
 )
+from argilla.server.schemas.v1.vector_settings import (
+    VectorSettings as VectorSettingsSchema,
+)
+from argilla.server.schemas.v1.vector_settings import (
+    VectorSettingsCreate,
+)
 from argilla.server.schemas.v1.vectors import Vector as VectorSchema
 from argilla.server.search_engine import SearchEngine
 from argilla.server.security.model import User
@@ -83,7 +86,6 @@ if TYPE_CHECKING:
 
     from argilla.server.schemas.v1.datasets import (
         DatasetUpdate,
-        VectorSettingsCreate,
     )
     from argilla.server.schemas.v1.fields import FieldUpdate
     from argilla.server.schemas.v1.questions import QuestionUpdate
